Change some contants' data type

+++ app/src/main/java/com/samalex/slucapstone/BoozymeterApplication.java
@@ -43,12 +43,12 @@ public class BoozymeterApplication extends Application {
     }
 
     // Official settings
-    final public static int CYCLE_OFFSET = 5 * 60 * 60 * 1000; // 5 hours (from start of natural cycle)
-    final public static int CYCLE_LENGTH = 24 * 60 * 60 * 1000; // 1 day
+    final public static long CYCLE_OFFSET = 5 * 60 * 60 * 1000; // 5 hours (from start of natural cycle)
+    final public static long CYCLE_LENGTH = 24 * 60 * 60 * 1000; // 1 day
     final public static int NUM_CYCLES = 21; // 21 cycles
-    final public static int SURVEY_OFFSET = 8 * 60 * 60 * 1000; // 8 hours (from start of natural cycle)
-    final public static int EVENING_REMINDER_OFFSET = 18 * 60 * 60 * 1000; // 18 hours (from start of natural cycle)
-    final public static int IN_EPISODE_REMINDER_INTERVAL = 30 * 60 * 1000; // 30 minutes
+    final public static long SURVEY_OFFSET = 8 * 60 * 60 * 1000; // 8 hours (from start of natural cycle)
+    final public static long EVENING_REMINDER_OFFSET = 18 * 60 * 60 * 1000; // 18 hours (from start of natural cycle)
+    final public static long IN_EPISODE_REMINDER_INTERVAL = 30 * 60 * 1000; // 30 minutes
     final public static int NUM_IN_EPISODE_REMINDER = 4; // 4 reminders every IN_EPISODE_REMINDER_INTERVAL
 
 //    // Testing settings
